+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes
 
 ### Unreleased
 
+* 2020-02-04 - Improvement: Support multilang strings in custom node titles - Thanks to Jean-Roch.
 * 2020-02-04 - Bugfix: A fatal error could have appeared during login as sessions - Credits to Adrian Perez.
 * 2019-09-04 - Fix flaws in the settings page regarding a link to other settings.